+++ tests/hermes_cli/test_setup_ollama_cloud_force_refresh.py
@@ -9,10 +9,13 @@ from __future__ import annotations
 
 def test_setup_ollama_cloud_passes_force_refresh(monkeypatch):
     """The provider-setup model-fetch for ollama-cloud must pass ``force_refresh=True``."""
-    import hermes_cli.main as main_mod
+    # The ollama-cloud branch lives in ``_model_flow_api_key_provider``, which was
+    # extracted from main.py into hermes_cli/model_setup_flows.py (god-file
+    # decomposition Phase 2). Inspect the module the code now lives in.
+    import hermes_cli.model_setup_flows as flows_mod
     import inspect
 
-    src = inspect.getsource(main_mod)
+    src = inspect.getsource(flows_mod)
 
     # Locate the ollama-cloud branch in the provider setup flow.
     marker = 'provider_id == "ollama-cloud"'
